kornbluh@widget0:~/c++/src> kornbluh@widget0:~/c++/src> sci objtools/format/feature_item.cpp
If a feature's gene xref points to a non-existent gene,
the old, incorrect behavior was to fall back on overlap.
Now, this has been fixed to instead leave the gene= line
as-is.  This also keeps compatibility with the C toolkit's
asn2gb program.

Example accession with gene xref problem: AF231993

Files touched:
feature_item.cpp

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@46979 78c7ea69-d796-4a43-9a09-de51944f1b03