CSeq_id: Defer warnings about unrecognized accession types.
When ignoring refinements, log the fact at level Info rather than
Warning, and hang on to the details.  If the caller turns out to be
interested in detailed classifications (as opposed to just being able
to construct Seq-ids from bare accessions) and looks up any accessions
with unrecognized preferred types, warn at that point, but even then
only the first time that situation occurs.

To that end, add a new CSeq_id::fParse_FallbackOK flag that callers
can supply to avoid triggering the warning when broad identification
is sufficient; use it internally as appropriate.

JIRA: CXX-9512.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@79972 78c7ea69-d796-4a43-9a09-de51944f1b03