Don't let FreeTDS's dbopen shadow the standard C library's.
On systems with a dbopen function in the standard C library, rename
FreeTDS's dbopen (historically left as is) after all.  We've gotten
away with this shadowing until now because our only use of FreeTDS's
dblib was via upstream unit tests, which have so far been unaffected.
However, on some operating systems (notably FreeBSD), looking up user
metadata such as the home directory can yield calls to dbopen, leading
to crashes when FreeTDS's implementation winds up receiving arguments
intended for the standard C library's.  An accidental omission in our
config.h adaptor (to be fixed shortly, now that it is safe to do so)
previously masked this problem by giving FreeTDS the impression that
the relevant lookup functions weren't available in the first place.
Preserve dbopen's historical exemption from renaming, at least for
now, when there is no dbopen in the standard library for it to shadow.

JIRA: CXX-9729.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@80639 78c7ea69-d796-4a43-9a09-de51944f1b03