Makefile.rules(_with_autodep).in: Honor new USE_FULL_PATHS macro.
When sources exist in the current directory, the build system
traditionally hasn't bothered supplying explicit paths to them.  Let
it optionally do so for the sake of code-scanning tools.

Because in-tree builds use separate build and source trees anyway,
this option mainly affects out-of-tree builds via import_project,
new_project, or the like.  These setups can now set this macro if
desired, and users can also supply it via the command line or
environment.

JIRA: CXX-9820.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@80915 78c7ea69-d796-4a43-9a09-de51944f1b03