CTL_LangCmd::x_GetDynamicID: Consider question mark count.
Don't even consider using ct_dynamic unless there are at least as many
question marks in the query string as supplied parameters.  It's
potentially okay to have extra question marks, because the extras may
be in strings, which this method doesn't attempt to parse out.
However, a shortage is a red flag and indicates that DBAPI should
stick with ct_command, even if some named parameters turn out not to
appear in the query string.  JIRA: CXX-5198.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@81057 78c7ea69-d796-4a43-9a09-de51944f1b03