CTL_Connection::x_GetNativeBlobDescriptor: Stop using TEXTPTR in WHERE.
The UPDATE statement intended to ensure TEXTPTR availability with a
minimum of side effects turned out to hang on Microsoft SQL Server in
some cases.  To avoid this problem, sanity check existing TEXTPTR
values in C++ rather than SQL and perform a second round trip as
necessary.  In the first pass, put the UPDATE statement behind an SQL
comment rather than omitting it altogether so that PubSeqOS still
knows to anticipate an incoming BLOB.  JIRA: CXX-9899.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@81290 78c7ea69-d796-4a43-9a09-de51944f1b03