CTL_Connection::x_GetNativeBlobDescriptor: Ensure XACT_ABORT is off.
Under Microsoft SQL Server, the (otherwise handled) failure to obtain
TEXTPTR values for columns that turned out to hold new-style BLOBs
(VARCHAR(MAX) or the like) could have the unintended side effect of
aborting the entire transaction stack when XACT_ABORT is on, as it is by
default nowadays (per CXX-6102).  Avoid this problem with the help of a
new internal RAII-style helper class, CAbortBlocker.  JIRA: GP-23437.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@81596 78c7ea69-d796-4a43-9a09-de51944f1b03