DBAPI: Clarify that BCP errors might not pertain to the latest row.
When performing a bulk insertion, arrange to keep track of how many rows
are in the current batch, and to make that count available to error-
reporting callbacks.  As such, if an error does occur in this context,
note that the cited row is simply the last one of the batch (passing
along its size) and might not be problematic itself.  Do so for both
client- and server-side errors; although the former are often immediate,
there are some cases in which they aren't.  JIRA: CXX-9911.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@82299 78c7ea69-d796-4a43-9a09-de51944f1b03