CDB_Numeric::x_MakeFromString: Tweak loop condition to avoid early termination.
In the loop that populates m_Body, bail only if p[i] not only starts
with 0 (historically the only condition) but also immediately follows
it up with a (terminal) 10; otherwise, this loop would bail
prematurely for nonzero values starting "0.0" or "-0.0".  Add an
assertion confirming that this tweak never results in running past the
start of m_Body (populated from back to front).

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@83117 78c7ea69-d796-4a43-9a09-de51944f1b03