CSeq_id: Support "eFasta_AsTypeAndContent" initialization.
Add constructor and Set variants that take the above special value
along with a (preparsed) choice value and a FASTA-style content string
with no leading tag (and vertical bars deeper in as appropriate).  To
that end, adjust x_Init not to expect a leading tag, and its callers
to strip them off as needed.  JIRA: CXX-10191.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@83360 78c7ea69-d796-4a43-9a09-de51944f1b03