SSeqIdRange: Support operator- between two iterators, missed earlier.
As with the Boolean comparison operators, assume that the argument is
legitimate to compare with *this.  (In all these cases, there's no
safety concern; the result just might not be meaningful.)  The debug
version of GNU libstdc++ (as enabled by --with-max-debug) caught this
operator variant's accidental absence.  JIRA: CXX-1916, RW-672.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@84483 78c7ea69-d796-4a43-9a09-de51944f1b03