CRPCClient: supply explicit Content-Type per data format.
When CRPCClient winds up issuing HTTP(S) requests (per either an explicit
URL or a named HTTP(S) service), it should arrange to send a Content-Type
header corresponding to the data format -- x-ncbi-data/x-asn-{text,binary}
or application/{xml,json}. As it is, CGIs or the like can otherwise
inappropriately attempt to parse POST-ed blobs as URL-encoded arguments,
which can misfire in various ways and in particular impede obtaining the
full raw blob. There are other protections that can and perhaps should be
in place here, but it's still best practice to set Content-Type explicitly.
JIRA: CXX-10758.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@86836 78c7ea69-d796-4a43-9a09-de51944f1b03