CID2CDDProcessor(_Impl): Synthesize errors for blob retrieval failures.
When unable to retrieve a blob from a satellite reserved for external
CDD annotations, synthesize an error rather than passing the buck to
other backends that aren't expected to be able to do anything useful
with the request.  If there was an exception along the way, log the
exception as a warning but don't pass it along to the client.  Make
the set of satellites covered formally configurable, to permit passing
the buck after all if this behavior turns out to be overzealous.
JIRA: CXX-10734.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@86870 78c7ea69-d796-4a43-9a09-de51944f1b03