JIRA: SMUPD-102
To CheckUpdateIds, add conditional handling for dead old GIs.
In regular mode, dead old GIs are flagged as "Dead_old_GI" errors (as opposed to "Wild dependency errors").
In alternative mode, dead old GIs will not be considered as errors at all.

The CHeckUpdateIds interface now provides for a new flag that specifies which of the two check modes will be 
used.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@87436 78c7ea69-d796-4a43-9a09-de51944f1b03