CDB_UserHandler_Deferred: Fully enforce max_severity; always use ERR_POST_X.
When catching an exception from PostMsg in Flush, account for the
possibility that it could somehow have a higher severity than the
exception passed to PostMsg.  Also, go ahead with ERR_POST_X here and
in the destructor, even when the only thing being posted is an
exception originating elsewhere.  JIRA: CXX-10938.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@87798 78c7ea69-d796-4a43-9a09-de51944f1b03