Implement a new CRPCClient<>::AsyncConnect method.
This method starts establishing a connection but does not fully wait
for it, and meanwhile lets the caller obtain the nascent connection's
OS-specific handle (file descriptor, on Unix) by passing in a pointer
(ignored if NULL) and length.  NB: Handle availability depends on
service type; handles should be available for standalone services like
ID2 and for stateful NCBID services, but not for HTTP(S) services,
which may have to make per-request connections.

To that end:
- Factor out much of x_Connect into a new x_GetConnNetInfo method,
  adding a little conditionalization so that x_ConnectURL (whose needs
  are somewhat more limited) can take advantage of this method too.
- Tweak x_Connect (called implicitly on demand) to follow up on
  AsyncConnect as needed.

JIRA: CXX-11275.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@89579 78c7ea69-d796-4a43-9a09-de51944f1b03