configure(.ac): Base GRPC*_LIBS on pkg-config --libs output if possible.
If we have pkg-config metadata for grpc++, use it as a general basis for
GRPC_LIBS.  This output makes a great starting point, complete with absl*
and upb as of the latest 1.28.1 release, but doesn't account for the
possibility of using deconflicted BoringSSL, and oddly misses Protobuf
altogether.  It also misses the need to use THREAD_LIBS explicitly in
single-threaded Toolkit builds, but that's more understandable.

Likewise for GRPC_UNSECURE_LIBS (from pkg-config grpc++_unsecure --libs),
apart from the latter's natural lack of any SSL dependency.

JIRA: CXX-11335.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@90002 78c7ea69-d796-4a43-9a09-de51944f1b03