Move new ID2 retriever framework to internal os_gateway tree.
Specifically, move it to .../os_gateway/plugins/api and put it in a
new osgapi library.  The framework is nominally general-use, but in
practice only anticipated to be of interest by OSG and its plugins,
which are themselves only of interest by OSG in practice, apart from a
couple of cases of gratuitous linkage.  The framework can move back if
and when some other project needs it after all.  Take the opportunity
to clean up spurious svn:executable properties.  JIRA: CXX-11051.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@90192 78c7ea69-d796-4a43-9a09-de51944f1b03