CGRPCRequestLogger: Actually define a constructor accepting streaming writers.
To that end, factor the existing constructor's body out into a new
x_Init method.  Take the opportunity to let the new variant accept
const streaming writers -- they're unlikely to be const in practice,
but might as well be given that this class doesn't use them at all.
JIRA: CXX-11379.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@90495 78c7ea69-d796-4a43-9a09-de51944f1b03