SAccGuide: Use a more scalable special-case representation when practical.
Introduce a new internal representation of special cases on top of
BitMagic's bvector<>, which should scale better to large numbers of
small blocks; automatically use it when feasible, but keep the current
representation as a fallback on an individual-block basis because it
supports arbitrarily large blocks.  (Support for a more compact file
format to follow separately when ready.)  JIRA: CXX-11813.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@93210 78c7ea69-d796-4a43-9a09-de51944f1b03