[IC]DBConnectionFactory: Populate an externally supplied exception container.
This ownership change is a prerequisite for narrower locking, barring
the otherwise gratuitous use of thread-local storage or the like.
Make the change explicit by reworking MakeDBConnection's interface
again -- it now takes a simple reference to an exception container to
use (which CDriverContext::MakeConnection allocates on the stack),
rather than a double pointer it's expected to populate.

CDBConnectionFactory: Follow through by incorporating this reference
into SOpeningContext and using it in lieu of the historical
factory-wide m_Errors member, now retired accordingly.  JIRA: CXX-4368.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@93268 78c7ea69-d796-4a43-9a09-de51944f1b03