tds_connect: Leave the original context's error handler alone.
Conditionalize out the code that temporarily set the socket's original
context's error handler to NULL during version probing.  The relevant
socket already uses a temporary copy of that context with a suitable
custom error handler, so the only potential usage of this context will
be from other threads that should be able to carry on independently.
JIRA: CXX-4368.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@93391 78c7ea69-d796-4a43-9a09-de51944f1b03