DBAPI CDriverContext: Accommodate multiple pool semaphore consumers.
To help ensure that it will be safe to lock less aggressively, introduce
a new inner struct SConsumer and maintain a list of instances thereof,
with MakePooledConnection now emplacing an entry at the end when it has
to wait and x_Recycle explicitly handing the connection off to the first
(longest waiting) matching instance, if any.  JIRA: CXX-4368.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@93405 78c7ea69-d796-4a43-9a09-de51944f1b03