check_make_*.sh, TestDriver.cmake: Forbid Boost.Test colorization.
Run tests with the environment variable BOOST_TEST_COLOR_OUTPUT set to
no.  This setting defaults to yes as of 1.68.0 (with no option for a
middle ground that would attempt to detect use of a terminal), and
should be safe to specify in this form even for old versions lacking
color support altogether.  JIRA: CXX-11801.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@93613 78c7ea69-d796-4a43-9a09-de51944f1b03