Allow safe asynchronous cancellation via CS_CANCEL_ATTN.
Under FreeTDS, stick with imposing a temporary short timeout when the
command to cancel is waiting for results -- if the server is truly
unresponsive, it might not even react to cancel packets.  At other
times, or with Sybase ctlib, track command state with the help of a
new RAII-style CTL_LRCmd::CCancelModeGuard and arrange to use the
async-safe CS_CANCEL_ATTN mode as appropriate, with other command
operations then automatically following up with full CS_CANCEL_ALL as
potentially needed.

To that end, have CTL_LRCmd additionally override MarkEndOfReply (to
clear the new m_CancelRequested flag) and require a cancellation type
(synchronous vs. asynchronous) for x_Cancel.  JIRA: CXX-11981.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@94114 78c7ea69-d796-4a43-9a09-de51944f1b03