accguide*: Broaden D????? classifications to gb_tpa_nuc, dropping _wgs.
Account for sharing the block with TSA and TLS records, which are
merely WGS-like.  (Yes, WGS has served as a fallback classification
for some TSA and TLS prefixes, but in those cases, rebuilding against
a fresher code base would yield a fully precise classification,
whereas in this case, the accessions would show up as WGS regardless.)
JIRA: ID-6954.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@95379 78c7ea69-d796-4a43-9a09-de51944f1b03