fix adjustment of cross origin alignments with cross-origin genomic gap; JIRA: PGAP-5677
Sometimes when we adjust cross origin alignments towards shrinking and when alignment has a cross-origin genomic gap we might encounter situation where we adjust the exon beyond the point where o
nly insertion part left from the exon.

Here we detect this situation and kill the exon. 

The question remains why did we keep the genomic_ins part in the cross origin alignment?
This might be a demand for cross origin alignment that a genome part seemlessly passes through origin without gaps
If this is the case, then we might have a situation in the future where after deleting one of the parts of the cross origin alignment we might end up with non-cross-origin alignment that has n in
sertion on the edge of the alignment, which is at least meaningless, but could be dangerous.

For now we are ignoring this possibility

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@95435 78c7ea69-d796-4a43-9a09-de51944f1b03