config.site.ncbi: Special-case the local path to libcurl.
Substitute the existing /opt/curl/lib for the customary (but not yet
present) /opt/ncbi/64/curl-7.85.0/*/lib*.  This adjustment makes for a
slight downgrade at the moment (to 7.84.0), but AFAICT should be fine
in practice.  In particular, I see no indication that anything uses
anything new to 7.85.0 (of which there isn't much) or insists on
having identical build-time and runtime versions.  JIRA: CXX-12668.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@98116 78c7ea69-d796-4a43-9a09-de51944f1b03