CDiagCollectGuard: Implement a middle-ground ePrintCapped action.
This action lets messages through but caps their severity at a
specified value, defaulting to the lowest severity collected.  Add
associated GetSeverityCap and SetSeverityCap methods, with the latter
accepting all levels (with the consequence of possibly releasing the
messages at too low a severity for them to appear after all).  Track
which thread-local post numbers are in range for a given guard, using
the resulting information to keep nested guards from adjusting the
severities of previously captured messages.  JIRA: CXX-12759.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@98621 78c7ea69-d796-4a43-9a09-de51944f1b03