accguide*: Add the latest round of interspersed TPE accessions.
This one includes both WGS nucleotide prefixes as usual (continuing
with CAOxxx) and the first new TPE proteins in a while, namely a large
crop of 3+7 TSA proteins (with CAH and CAI prefixes).  The latter is
v2 only, but v1 is of increasingly limited interest anyway.

Update the protein high-water-mark comment.

JIRA: CXX-12816.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@99042 78c7ea69-d796-4a43-9a09-de51944f1b03