CPoolOfThreads: Move away from volatile per C++20 deprecations.
Many uses of volatile are now deprecated because it was always mainly
intended for intrathread coordination with signal handlers or the
(exotic) like.  Conservatively substitute CAtomicCounter_WithAutoInit
for m_Max(Urgent)Threads and atomic<int> for m_Delta, which needs to
be able to take on negative values whereas CAtomicCounter_WithAutoInit
may be unsigned.  JIRA: CXX-12877.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@99264 78c7ea69-d796-4a43-9a09-de51944f1b03