Support having $(libdir) in (the original) LD_LIBRARY_PATH.
When arranging to reduce LD_LIBRARY_PATH to a single directory
populated by symlinks, account for the possibility that the original
value contained this build's $(libdir), in which case building shared
libraries should yield corresponding symlinks from that single
directory to $(libdir).  Such LD_LIBRARY_PATH settings are notably
helpful in conjunction with --with(-relative)-runpath=... flags that
don't supply build-time paths to this build's libraries.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@99338 78c7ea69-d796-4a43-9a09-de51944f1b03