CDBConnectionFactory::CServiceInfo::GetOptions: Refresh if expired.
If any options are past their expiration times, obtain a fresh list of
options and swap it in if it's not empty.  (If the fresh list is
somehow empty, keep using the existing list until either a non-empty
list shows up [checking every time at this point] or something leads
to forcibly clearing it.)  JIRA: CXX-12884.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@100186 78c7ea69-d796-4a43-9a09-de51944f1b03