Fix builds with pre-1.73 Boost.Test in C++20 mode.
In such environments (e.g., Ubuntu 20.04 [focal] with a GCC 11
backport), the wchar_t const* variant of print_log_value tries to use
a corresponding deleted variant of operator<<.  The good news is that
this deletion is via a template and as such possible to rescind by
defining a non-template variant specific to std::ostream (sufficient
in practice) in namespace std in the handful of places that need it.
JIRA: CXX-12934.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@101567 78c7ea69-d796-4a43-9a09-de51944f1b03