bulk.c 1.4.x (tds_bcp_send_record): Reject invalid NULL data.
When the user erroneously supplies NULL data for a column which does not
use a nullable type and is not otherwise known to be nullable, cleanly
fail.  Pressing on anyway (in hopes that a server-side default would
take effect?) was liable to trigger legitimate assertion failures in
tds_checks.c.  JIRA: CXX-13176.  TRUNK-1.00: 554190.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@102173 78c7ea69-d796-4a43-9a09-de51944f1b03