Check for OpenTelemetry (typically using version 1.14.2).
Pull in the otlp/http exporter for mainstream use and the ostream
exporter to help with any needed debugging, both in trace/span
flavors.

On the Conan front, leave the key self.requires call commented out
until we can rely on Conan 1.60+ (as required by the current libtool
recipe, pulled in for the sake of libcurl), but go ahead with
requisite tuneups.  In particular, explicitly use Protobuf 3.21.12
(gRPC 1.50.1 merely requires 3.21.4, and Conan doesn't reconcile the
requirements on its own) and request static builds of googleapis/* and
grpc-proto/* in keeping with the general preference for static builds.

JIRA: CXX-13550.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@102439 78c7ea69-d796-4a43-9a09-de51944f1b03