ct_describe 1.4.x: Correctly complain about unsupported server types.
The previous change wrongly issued this error (only) in a different
scenario (out-of-range item number), to which end it performed a
potentially out-of-range array access.  Switch the order of the
access and that check back and issue this error prior to the CS_FAIL
return to which it should actually correspond.  JIRA: CXX-13176.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@102460 78c7ea69-d796-4a43-9a09-de51944f1b03