Expose (S)DBAPI drivers' underlying library versions.
Expose the version as a raw string from the driver layer, a (cached)
CVersionInfo from the user layer, and an EDriver value from SDBAPI.
Formally do so on a per-connection basis to accommodate ODBC, for
which the backend version is of much more interest than the driver
manager's and needs to be obtained from a connection because the
underlying API permits using different backends for different
connections.  JIRA: CXX-13629.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@102711 78c7ea69-d796-4a43-9a09-de51944f1b03