python-config.py: Favor sysconfig over old distutils.sysconfig.
Python 3.12 dropped distutils from the standard library; some 3.12+
installations have it grafted back on, but we shouldn't count on that.
To that end, rely only on get_config_var and get_config_vars, and in
particular NOT on get_python_inc.  Fall back to distutils.sysconfig if
necessary to retain compatibility with ancient Python versions.
JIRA: CXX-13780.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@104485 78c7ea69-d796-4a43-9a09-de51944f1b03