CGRPCServerCallbacks: URL-decode peer values.  JIRA: CGV-1201.
gRPC presents peers as nominal URIs with schemes like ipv4 or ipv6.
In the latter case, because it supplies a port number, gRPC wraps the
IP address in square brackets per standard colon-disambiguation
practice; as of version 1.47-pre1, it proceeds to encode the brackets
as %5B and %5D for strict RFC 3986 compliance, making it necessary to
account for that possibility.

NB: x_IsRealRequest decodes the whole URI for simplicity rather than
splitting it first; the latter would be formally better practice but
would make no actual difference.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@105462 78c7ea69-d796-4a43-9a09-de51944f1b03